[crypto] Fix P384 ECDSA remasking typos
Amends a typo in the generation of n << 64 in `p384_scalar_mask`, which
otherwise causes the generated mask to not be uniform modulo n << 64.
Additionally fixes a small typo in `p384_scalar_mask_test` which causes
the loaded value of n << 64 to be loaded as zero, which causes the
test to still be functionally correct but doesn't match the intended
behavior per the comments.

Signed-off-by: Kat Fox <kat@zerorisc.com>