[csrng/rtl] Remove the pdata FIFO from ctr_drbg_upd
This commit removes the pdata FIFO from the csrng_ctr_drbg_upd stage. As
a consquence, pdata for the final step is now taken from the request
fifo right at the request port of this module. This requires the
internal control logic to wait for the block_encrypt module to process
all three requests until the upstream req_rdy_o can be asserted. This
change, however, does not reduce performance/throughput of the CSRNG.

Signed-off-by: Florian Glaser <glaserf@lowrisc.org>